+++ task-manager/src/index.js
@@ -1,6 +1,7 @@
 const express = require('express')
 require('./db/mongoose')
 const User = require('./models/user')
+const Task = require('./models/task')
 
 const app = express()
 const port = process.env.PORT || 3000
@@ -16,8 +17,17 @@ app.post('/users', (req, res) => {
         console.log(error)
         res.status('400').send(error)
     })
-
 })
+
+app.post('/tasks', (req, res) => {
+    const task = new Task(req.body)
+    task.save().then((response) => {
+        res.send(response)
+    }).catch((error) => {
+        res.status("400").send(error)
+    })
+})
+
 
 app.listen(port, () => {
     console.log(`Server is up on port: ${port}.`)
